fix: refactor account component and added test components
